rename store folder
