+++ packages/js/product-editor/src/store/product-editor-ui/reducer.ts
